ModalExport: Refactor advancedSettingsArray.
